Moved some tests out of the ExpressionParser test
